Bug #1401225 bsp: camera: Change for pclint
[root cause  ] Improper variable type used
[changes     ] isp2.6
[side effects]NA
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums512_1h10
[test case]:camera funtion
[reviewers   ] yangyang.liu
[change_type ] feature
[tag_product ] common
Improper variable type used.

+++ common/cam_trusty.h
@@ -129,7 +129,7 @@ bool cam_trusty_isp_fetch_addr_set(unsigned long y_addr,
 	unsigned long u_addr, unsigned long v_addr);
 bool  cam_trusty_isp_pitch_set(uint32_t y_pitch, uint32_t u_pitch,
 	uint32_t v_pitch);
-bool  cam_trusty_isp_3dnr_fetch_set(uint32_t chroma, uint32_t luma,
+bool  cam_trusty_isp_3dnr_fetch_set(unsigned long chroma, uint32_t luma,
 	uint32_t pitch);
 bool  cam_trusty_csi_switch_ctrl_set(uint32_t csi_sel_ctrl);
 bool  cam_trusty_security_set(struct  sprd_cam_sec_cfg *camsec_cfg,
